requirements.txt
httpx==0.23.0
apscheduler==3.9.1
pycryptodome==3.15.0